working on cms cruds

+++ app/Http/Controllers/CmsController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\UpdateCmsRequest;
 use App\Http\Controllers\AppBaseController;
 use App\Repositories\CmsRepository;
 use Illuminate\Http\Request;
+use App\Models\Cms;
 use Flash;
 
 class CmsController extends AppBaseController
@@ -32,17 +33,29 @@ class CmsController extends AppBaseController
      */
     public function create()
     {
-        return view('cms.create');
+        $pages = Cms::pluck('title', 'id');
+        return view('cms.create', compact('pages'));
     }
 
     /**
      * Store a newly created Cms in storage.
      */
     public function store(CreateCmsRequest $request)
-    {
-        $input = $request->all();
+    {   
+        $fieldsToUpdate = $request->except('gallery');
+        $cms = $this->cmsRepository->create($fieldsToUpdate);
+        if ($request->hasfile('banner_image')) {
+            $cms->banner_image = uploadImage($request->file('banner_image'), CMS_IMAGE_PATH);
+        }
+        $cms->main_menu = $request->has('main_menu') ? 1 : 0;
+        $cms->top_menu = $request->has('top_menu') ? 1 : 0;
+        $cms->side_menu = $request->has('side_menu') ? 1 : 0;
+        $cms->footer_menu = $request->has('footer_menu') ? 1 : 0;
 
-        $cms = $this->cmsRepository->create($input);
+        $cms->gallery = uploadMultipleImage($request->file('gallery'), CMS_IMAGE_PATH, $request->multiple_alt_textgallery, null);
+        $cms->save();
+
+        
 
         Flash::success('Cms saved successfully.');
 
@@ -77,8 +90,9 @@ class CmsController extends AppBaseController
 
             return redirect(route('cms.index'));
         }
+        $pages = Cms::pluck('title', 'id');
 
-        return view('cms.edit')->with('cms', $cms);
+        return view('cms.edit',compact('pages'))->with('cms', $cms);
     }
 
     /**
@@ -94,11 +108,26 @@ class CmsController extends AppBaseController
             return redirect(route('cms.index'));
         }
 
-        $cms = $this->cmsRepository->update($request->all(), $id);
+        if ($request->hasfile('banner_image')) {
+            removeImage($cms->banner_image, CMS_IMAGE_PATH);
+        }
+        $fieldsToUpdate = $request->except('gallery');
+        $cms = $this->cmsRepository->update($fieldsToUpdate, $id);
+        if ($request->hasfile('banner_image')) {
+            $cms->banner_image = uploadImage($request->file('banner_image'), CMS_IMAGE_PATH);
+        }
+        $cms->main_menu = $request->has('main_menu') ? 1 : 0;
+        $cms->top_menu = $request->has('top_menu') ? 1 : 0;
+        $cms->side_menu = $request->has('side_menu') ? 1 : 0;
+        $cms->footer_menu = $request->has('footer_menu') ? 1 : 0;
+        $cms->gallery = uploadMultipleImage($request->file('gallery'), CMS_IMAGE_PATH, $request->multiple_alt_textgallery, $cms->gallery);
+        $cms->save();
 
+   
         Flash::success('Cms updated successfully.');
 
         return redirect(route('cms.index'));
+
     }
 
     /**
@@ -108,6 +137,7 @@ class CmsController extends AppBaseController
      */
     public function destroy($id)
     {
+        
         $cms = $this->cmsRepository->find($id);
 
         if (empty($cms)) {
@@ -116,10 +146,40 @@ class CmsController extends AppBaseController
             return redirect(route('cms.index'));
         }
 
-        $this->cmsRepository->delete($id);
+        if (getSubMenu($cms->id)->count() > 0) {
+            Flash::error('Unable to delete because this page contains sub pages.');
+        } else {
+            if ($cms->banner_image != '') {
+                removeImage($cms->banner_image, CMS_IMAGE_PATH);
+            }
 
-        Flash::success('Cms deleted successfully.');
+            if ($cms->gallery != '') {
+                foreach (json_decode($cms->gallery, true) as $gal) {
+                    removeImage($gal['path'], CMS_IMAGE_PATH);
+                }
+            }
+
+            $this->cmsRepository->delete($id);
+
+            Flash::success('Cms deleted successfully.');
+        }
 
         return redirect(route('cms.index'));
+    }
+
+    public function removeGalleryItem($id, $key)
+    {
+        $cms = $this->cmsRepository->find($id);
+        if (!empty($cms)) {
+            $data = json_decode($cms->gallery, true);
+            removeImage($data[$key]['path'], CMS_IMAGE_PATH);
+            unset($data[$key]);
+            $cms->gallery = json_encode(array_values($data));
+            $cms->save();
+            Flash::success('Image Removed Successfully.');
+            return redirect()->back();
+        } else {
+            return redirect()->back();
+        }
     }
 }
